% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLTuniform.R
\name{mycltu}
\alias{mycltu}
\title{CLT Uniform}
\usage{
mycltu(n = 1, iter = 10000, a = 0, b = 10, x = NULL)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{Number of Iterations to generate random samples}

\item{a}{The lower limit of the uniform distribution}

\item{b}{The upper limit of the uniform distribution}

\item{x}{Optional vector of x values for plotting.}
}
\value{
Generate histograms and density plots for the Central Limit Theorem
}
\description{
CLT Uniform
}
\examples{
mycltu(n=1,iter=1000,a=0,b=10)
}
