% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Max Likelihood with Two Parameters}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{A vector containing the sample data.}

\item{mu}{A vector of possible mean values to evaluate.}

\item{sig}{A vector of possible standard deviation values to evaluate.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
A list containing the sample data (x), the coordinates of the maximum likelihood estimate (coord), and the maximum likelihood value (maxl).
}
\description{
Max Likelihood with Two Parameters
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)

}
