% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot Normal Distribution Curve with Shaded Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution.}

\item{a}{The upper limit of integration}
}
\value{
A graph of the normal distribution
}
\description{
Plot Normal Distribution Curve with Shaded Area
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1)
}
