% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{myPlot}
\alias{myPlot}
\title{Title Quadratic Function for calculating Spruce Height}
\usage{
myPlot(x)
}
\arguments{
\item{x}{BHDiameter of Spruce}
}
\value{
Height of Spruce
}
\description{
Title Quadratic Function for calculating Spruce Height
}
\examples{
myPlot(x = 1:10)
}
