% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePlot.R
\name{mysample}
\alias{mysample}
\title{Title Generate Barplots of Random Samples}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{The number of samples to draw from the set}

\item{iter}{The number of iterations to repeat the sampling process. Default is 10.}

\item{time}{The time in seconds to pause between generating each barplot. Default is 0.5.}
}
\value{
Generates barplots
}
\description{
Title Generate Barplots of Random Samples
}
\examples{
mysample(n = 20, iter = 10, time = 1)
}
