% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Objective vs n to find optimal tickets sold}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats in the flight}

\item{gamma}{Probability a plane will be overbooked}

\item{p}{Probability of a "show"}
}
\value{
A named list of nd, nc, N, gamma, and p as well as a continuous and discrete plot.
}
\description{
Objective vs n to find optimal tickets sold
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
